third_party/blink/web_tests/external/wpt/css/css-animations/parsing/keyframes-name-invalid-expected.txt

This is a testharness.js-based test.
[FAIL] invalid: @keyframes none { }
  assert_equals: expected 0 but got 1
[FAIL] invalid: @keyframes default { }
  assert_equals: expected 0 but got 1
[FAIL] invalid: @keyframes initial { }
  assert_equals: expected 0 but got 1
[FAIL] invalid: @keyframes inherit { }
  assert_equals: expected 0 but got 1
[FAIL] invalid: @keyframes unset { }
  assert_equals: expected 0 but got 1
[FAIL] invalid: @keyframes revert { }
  assert_equals: expected 0 but got 1
[FAIL] invalid: @keyframes revert-layer { }
  assert_equals: expected 0 but got 1
Harness: the test ran to completion.
